+++ Mohashan.UserManager.Application/Features/Group/Commands/DeleteGroup/DeleteGroupCommandHandler.cs
@@ -1,18 +1,24 @@
 ﻿using MediatR;
 using Mohashan.UserManager.Application.Contracts.Persistence;
+using Mohashan.UserManager.Application.Features.Feature.Commands.DeleteFeature;
 
 namespace Mohashan.UserManager.Application.Features.Group.Commands.DeleteGroup;
 
 public class DeleteGroupCommandHandler : IRequestHandler<DeleteGroupCommand>
 {
-    private readonly IAsyncRepository<Domain.Entities.Group> _groupRepository;
+    private readonly IGroupRepository _groupRepository;
 
-    public DeleteGroupCommandHandler(IAsyncRepository<Domain.Entities.Group> groupRepository)
+    public DeleteGroupCommandHandler(IGroupRepository groupRepository)
     {
         _groupRepository = groupRepository;
     }
     public async Task<Unit> Handle(DeleteGroupCommand request, CancellationToken cancellationToken)
     {
+        var validator = new DeleteGroupCommandValidator(_groupRepository);
+        var validatorResult = await validator.ValidateAsync(request);
+        if (!validatorResult.IsValid)
+            throw new Exceptions.ValidationException(validatorResult);
+
         await _groupRepository.DeleteAsync(request.Id);
         return Unit.Value;
     }
