+++ Mohashan.UserManager.Application/Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using Mohashan.UserManager.Application.Contracts.Persistence;
 using Mohashan.UserManager.Domain.Entities;
+using System.Collections.Immutable;
 
 namespace Mohashan.UserManager.Application.Features.Users.Commands.DeleteUser;
 
@@ -14,6 +15,11 @@ public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand>
     }
     public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
+        var validator = new DeleteUserCommandValidator(_userRepository);
+        var validatorResult = await validator.ValidateAsync(request);
+        if (!validatorResult.IsValid)
+            throw new Exceptions.ValidationException(validatorResult);
+
         await _userRepository.DeleteAsync(request.Id);
         return Unit.Value;
     }
